requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
autopep8==2.3.2
cfgv==3.4.0
decorator==5.2.1
distlib==0.3.9
Django==5.1.6
django-auto-prefetch==1.12.0
django-extensions==3.2.3
django-rest-framework==0.1.0
django-silk==5.3.2
djangorestframework==3.15.2
drf-spectacular==0.28.0
executing==2.2.0
Faker==36.1.1
filelock==3.17.0
gprof2dot==2024.6.6
gunicorn==23.0.0
identify==2.6.8
inflection==0.5.1
ipython==9.0.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matplotlib-inline==0.1.7
nodeenv==1.9.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.13.0
Pygments==2.19.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
virtualenv==20.29.2
wcwidth==0.2.13
